oscillators

+++ src/notes.h
@@ -1,3 +1,3 @@
-extern uint32_t note[128];
+extern float note[128];
 
 void gen_note_table(void);
